day 18 part1

+++ src/main/kotlin/y2024/day17/Parser.kt
@@ -3,16 +3,16 @@ package y2024.day17
 import utils.split
 
 object Parser {
-    data class Input(var regA: Int, var regB: Int, var regC: Int, val output: MutableList<Int>, val program: List<Int>)
+    data class Input(var regA: Long, var regB: Long, var regC: Long, val output: MutableList<Long>, val program: List<Long>)
     fun parse(data: List<String>): Input {
         val (registers, program) = data.split { it.isBlank() }
 
         return Input(
-            registers[0].split(": ").last().toInt(),
-            registers[1].split(": ").last().toInt(),
-            registers[2].split(": ").last().toInt(),
+            registers[0].split(": ").last().toLong(),
+            registers[1].split(": ").last().toLong(),
+            registers[2].split(": ").last().toLong(),
             mutableListOf(),
-            program.first().split(": ").last().split(",").map { it.toInt() }
+            program.first().split(": ").last().split(",").map { it.toLong() }
         )
     }
 }
